edgeai-benchmark/requirements.txt
cython
wheel
pillow
pycocotools
colorama
tqdm
pyyaml
pytest
json_tricks
munkres
scipy
wurlitzer
plyfile
scikit-learn
debugpy
